+++ nautilus_trader/adapters/betfair/providers.pyx
@@ -255,7 +255,8 @@ def make_instruments(market_definition, currency):
             selection_handicap=str(runner.get("hc", runner.get("handicap")) or "0.0"),
             currency=currency,
             # TODO - Add the provider, use clock
-            timestamp_ns=unix_timestamp_ns()
+            timestamp_origin_ns=unix_timestamp_ns(),  # TODO(bm): Duplicate timestamps for now
+            timestamp_ns=unix_timestamp_ns(),
             # info=market_definition,  # TODO We should probably store a copy of the raw input data
         )
         instruments.append(instrument)