+++ nautilus_trader/adapters/ccxt/data.pyx
@@ -689,7 +689,8 @@ cdef class CCXTDataClient(LiveMarketDataClient):
                         level=level,
                         bids=list(bids),
                         asks=list(asks),
-                        timestamp_ns=self._ccxt_to_timestamp_ns(millis=timestamp_ms)
+                        timestamp_origin_ns=self._ccxt_to_timestamp_ns(millis=timestamp_ms),
+                        timestamp_ns=self._clock.timestamp_ns(),
                     )
 
                     self._handle_data(snapshot)
@@ -769,6 +770,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
                     best_bid[1],
                     best_ask[1],
                     timestamp,
+                    self._clock.timestamp_ns(),
                     price_precision,
                     size_precision,
                 )
@@ -787,6 +789,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
         double best_ask,
         double best_bid_size,
         double best_ask_size,
+        int64_t timestamp_origin_ns,
         int64_t timestamp_ns,
         int price_precision,
         int size_precision,
@@ -797,6 +800,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
             Price(best_ask, price_precision),
             Quantity(best_bid_size, size_precision),
             Quantity(best_ask_size, size_precision),
+            timestamp_origin_ns,
             timestamp_ns,
         )
 
@@ -833,6 +837,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
                     trade["side"],
                     trade["id"],
                     self._ccxt_to_timestamp_ns(millis=trade["timestamp"]),
+                    self._clock.timestamp_ns(),
                     price_precision,
                     size_precision,
                 )
@@ -851,6 +856,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
         double amount,
         str aggressor_side,
         str trade_match_id,
+        int64_t timestamp_origin_ns,
         int64_t timestamp_ns,
         int price_precision,
         int size_precision,
@@ -861,6 +867,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
             Quantity(amount, size_precision),
             AggressorSideParser.from_str(aggressor_side.upper()) ,
             TradeMatchId(trade_match_id),
+            timestamp_origin_ns,
             timestamp_ns,
         )
 
@@ -921,6 +928,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
                         bar[4],
                         bar[5],
                         this_timestamp,
+                        self._clock.timestamp_ns(),
                         price_precision,
                         size_precision,
                     )
@@ -940,6 +948,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
         double low_price,
         double close_price,
         double volume,
+        int64_t timestamp_origin_ns,
         int64_t timestamp_ns,
         int price_precision,
         int size_precision,
@@ -951,6 +960,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
             Price(low_price, price_precision),
             Price(close_price, price_precision),
             Quantity(volume, size_precision),
+            timestamp_origin_ns,
             timestamp_ns,
         )
 
@@ -1158,6 +1168,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
             AggressorSide.BUY if trade["side"] == "buy" else AggressorSide.SELL,
             TradeMatchId(trade["id"]),
             self._ccxt_to_timestamp_ns(millis=trade["timestamp"]),
+            self._clock.timestamp_ns(),
         )
 
     cdef inline Bar _parse_bar(
@@ -1175,6 +1186,7 @@ cdef class CCXTDataClient(LiveMarketDataClient):
             Price(values[4], price_precision),
             Quantity(values[5], size_precision),
             self._ccxt_to_timestamp_ns(millis=values[0]),
+            self._clock.timestamp_ns(),
         )
 
     cdef str _make_timeframe(self, BarSpecification bar_spec):