+++ nautilus_trader/common/clock.pyx
@@ -41,7 +41,7 @@ cdef class Clock:
     """
     The abstract base class for all clocks.
 
-    This class should not be used directly, but through its concrete subclasses.
+    This class should not be used directly, but through a concrete subclass.
     """
 
     def __init__(self):