+++ tests/test_kit/providers.py
@@ -149,8 +149,13 @@ class TestDataProvider:
                         size=Quantity(d["trade"]["volume"], 4),
                         aggressor_side=d["trade"]["side"],
                         match_id=TradeMatchId(d["trade"]["trade_id"]),
-                        timestamp_ns=millis_to_nanos(
+                        timestamp_origin_ns=millis_to_nanos(
                             pd.Timestamp(d["remote_timestamp"]).timestamp()
+                        ),
+                        timestamp_ns=millis_to_nanos(
+                            pd.Timestamp(
+                                d["remote_timestamp"]
+                            ).timestamp()  # TODO(cs): Hardcoded identical for now
                         ),
                     ),
                 }
@@ -283,6 +288,7 @@ class TestInstrumentProvider:
             margin_maint=Decimal(),
             maker_fee=Decimal("0.001"),
             taker_fee=Decimal("0.001"),
+            timestamp_origin_ns=0,
             timestamp_ns=0,
         )
 
@@ -324,6 +330,7 @@ class TestInstrumentProvider:
             margin_maint=Decimal("0.35"),
             maker_fee=Decimal("0.0001"),
             taker_fee=Decimal("0.0001"),
+            timestamp_origin_ns=0,
             timestamp_ns=0,
         )
 
@@ -365,6 +372,7 @@ class TestInstrumentProvider:
             margin_maint=Decimal("0.0035"),
             maker_fee=Decimal("-0.00025"),
             taker_fee=Decimal("0.00075"),
+            timestamp_origin_ns=0,
             timestamp_ns=0,
         )
 
@@ -406,6 +414,7 @@ class TestInstrumentProvider:
             margin_maint=Decimal("0.007"),
             maker_fee=Decimal("-0.00025"),
             taker_fee=Decimal("0.00075"),
+            timestamp_origin_ns=0,
             timestamp_ns=0,
         )
 
@@ -473,5 +482,6 @@ class TestInstrumentProvider:
             margin_maint=Decimal("0.03"),
             maker_fee=Decimal("0.00002"),
             taker_fee=Decimal("0.00002"),
+            timestamp_origin_ns=0,
             timestamp_ns=0,
         )